Fixes rebasing issues with favourites
